feat: add more tests and features

+++ zio-spark-core/src/test/scala/zio/spark/helper/Fixture.scala
@@ -6,7 +6,7 @@ import zio.spark.sql.TryAnalysis.syntax.throwAnalysisException
 import zio.spark.sql.implicits._
 
 object Fixture {
-  final case class Person(name: String, age: Int)
+  final case class Person(name: String, age: Long)
 
   def readCsv(path: String): SIO[DataFrame] = SparkSession.read.inferSchema.withHeader.withDelimiter(";").csv(path)
